requirements.txt
coreapi==2.2.4
coreschema==0.0.4
dj-database-url==0.4.2
Django==1.10.5
django-appconf==1.0.2
django-imagekit==4.0
django-rest-swagger==2.1.1
djangorestframework==3.5.4
itypes==1.1.0
Jinja2==2.9.5
MarkupSafe==1.0
olefile==0.44
openapi-codec==1.3.1
pilkit==2.0
Pillow==4.0.0
Pygments==2.2.0
python-decouple==3.0
requests==2.13.0
simplejson==3.10.0
six==1.10.0
Unipath==1.1
uritemplate==3.0.0
